Fixed issue 279
