+++ test/functionalTest/cases/0000_ngsild/ngsild_add_attr.test
@@ -186,17 +186,17 @@ MongoDB server version: REGEX(.*)
 		},
 		"https://uri=etsi=org/ngsi-ld/default-context/P2" : {
 			"type" : "Property",
-			"value" : "new",
-			"mdNames" : [ ],
 			"creDate" : REGEX(.*),
-			"modDate" : REGEX(.*)
+			"modDate" : REGEX(.*),
+			"value" : "new",
+			"mdNames" : [ ]
 		},
 		"https://uri=etsi=org/ngsi-ld/default-context/P3" : {
 			"type" : "Property",
-			"value" : false,
-			"mdNames" : [ ],
 			"creDate" : REGEX(.*),
-			"modDate" : REGEX(.*)
+			"modDate" : REGEX(.*),
+			"value" : false,
+			"mdNames" : [ ]
 		}
 	},
 	"creDate" : REGEX(.*),
