+++ test/functionalTest/cases/0000_ngsild/ngsild_adding_attributes.test
@@ -239,44 +239,34 @@ MongoDB server version: REGEX(.*)
 		},
 		"https://fiware=github=io/tutorials=Step-by-Step/schema/stocks" : {
 			"type" : "Relationship",
-			"value" : "urn:ngsi-ld:Product:001",
-			"mdNames" : [ ],
 			"creDate" : REGEX(.*),
-			"modDate" : REGEX(.*)
+			"modDate" : REGEX(.*),
+			"value" : "urn:ngsi-ld:Product:001",
+			"mdNames" : [ ]
 		},
 		"https://fiware=github=io/tutorials=Step-by-Step/schema/locatedIn" : {
 			"type" : "Relationship",
+			"creDate" : REGEX(.*),
+			"modDate" : REGEX(.*),
 			"value" : "urn:ngsi-ld:Building:store001",
-			"mdNames" : [
-				"https://fiware.github.io/tutorials.Step-by-Step/schema/installedBy",
-				"https://fiware.github.io/tutorials.Step-by-Step/schema/statusOfWork"
-			],
 			"md" : {
-				"https://fiware.github.io/tutorials.Step-by-Step/schema/installedBy" : {
+				"https://fiware=github=io/tutorials=Step-by-Step/schema/installedBy" : {
 					"type" : "Relationship",
 					"value" : "urn:ngsi-ld:Person:employee001"
 				},
-				"https://fiware.github.io/tutorials.Step-by-Step/schema/statusOfWork" : {
+				"https://fiware=github=io/tutorials=Step-by-Step/schema/statusOfWork" : {
 					"type" : "Property",
 					"value" : "https://fiware.github.io/tutorials.Step-by-Step/schema/completed"
 				}
 			},
-			"creDate" : REGEX(.*),
-			"modDate" : REGEX(.*)
+			"mdNames" : [
+				"https://fiware.github.io/tutorials.Step-by-Step/schema/installedBy",
+				"https://fiware.github.io/tutorials.Step-by-Step/schema/statusOfWork"
+			]
 		}
 	},
 	"creDate" : REGEX(.*),
 	"modDate" : REGEX(.*),
-	"location" : {
-		"attrName" : "location",
-		"coords" : {
-			"type" : "Point",
-			"coordinates" : [
-				13.3986112,
-				52.554699
-			]
-		}
-	},
 	"lastCorrelator" : ""
 }
 bye
