+++ test/functionalTest/cases/0000_ngsild/ngsild_testsuite-simple_subscription_to_concrete_attribute.test
@@ -37,8 +37,8 @@ accumulatorStart --pretty-print
 # 01. Creation of entity
 # 02. Creation of subscription - initial notification should be received by the accumulator
 # 03. Dump and reset accumulator. See one notification
-# 04. Update the entity - provoking a notification - except ... notifications for POST /entities/attrs without 'options=noOverwrite' hasn't been implemented
-# 05. Dump accumulator. See no notifications
+# 04. Update the entity - provoking a notification
+# 05. Dump accumulator. See the notification
 #
 
 echo "01. Creation of entity"
@@ -91,8 +91,8 @@ echo
 echo
 
 
-echo "04. Update the entity - provoking a notification - except ... notifications for POST /entities/attrs without 'options=noOverwrite' hasn't been implemented"
-echo "=========================================================================================================================================================="
+echo "04. Update the entity - provoking a notification"
+echo "================================================"
 payload='{
   "type": "Property",
   "value": 5
@@ -103,7 +103,7 @@ echo
 
 
 sleep 1
-echo "05. Dump accumulator. See no notifications"
+echo "05. Dump accumulator. See the notification"
 echo "=========================================="
 accumulatorDump
 echo
@@ -166,8 +166,8 @@ Content-Type: application/json; charset=utf-8
 =======================================
 
 
-04. Update the entity - provoking a notification - except ... notifications for POST /entities/attrs without 'options=noOverwrite' hasn't been implemented
-==========================================================================================================================================================
+04. Update the entity - provoking a notification
+================================================
 HTTP/1.1 204 No Content
 Content-Length: 0
 Link: <https://uri.etsi.org/ngsi-ld/v1/ngsi-ld-core-context.jsonld>; rel="http://www.w3.org/ns/json-ld#context"; type="application/ld+json"
@@ -175,8 +175,39 @@ Date: REGEX(.*)
 
 
 
-05. Dump accumulator. See no notifications
+05. Dump accumulator. See the notification
 ==========================================
+POST http://REGEX(.*)
+Fiware-Servicepath: /
+Content-Length: 346
+User-Agent: REGEX(.*)
+Ngsiv2-Attrsformat: normalized
+Host: localhost:REGEX(.*)
+Accept: application/json
+Link: <https://uri.etsi.org/ngsi-ld/v1/ngsi-ld-core-context.jsonld>; rel="http://www.w3.org/ns/json-ld#context"; type="application/ld+json"
+Content-Type: application/json; charset=utf-8
+
+{
+    "data": [
+        {
+            "brandName": {
+                "type": "Property", 
+                "value": "Mercedes"
+            }, 
+            "id": "urn:ngsi-ld:Vehicle:V1234", 
+            "speed": {
+                "type": "Property", 
+                "value": 34
+            }, 
+            "type": "Vehicle"
+        }
+    ], 
+    "id": "urn:ngsi-ld:Notification:REGEX(.*)", 
+    "notifiedAt": "REGEX(.*)", 
+    "subscriptionId": "urn:ngsi-ld:Subscription:mySubscription:20-09-2019:13:13:00", 
+    "type": "Notification"
+}
+=======================================
 
 
 --TEARDOWN--
